+++ examples/main/shared/viewer/Viewer.jsx
@@ -5,8 +5,8 @@ import { RichContentModal, isSSR } from 'wix-rich-content-common';
 import * as PropTypes from 'prop-types';
 import * as Plugins from './ViewerPlugins';
 import theme from '../theme/theme'; // must import after custom styles
-// import getImagesData from 'wix-rich-content-fullscreen/src/lib/getImagesData';
-// import Fullscreen from 'wix-rich-content-fullscreen';
+import getImagesData from 'wix-rich-content-fullscreen/src/lib/getImagesData';
+import Fullscreen from 'wix-rich-content-fullscreen';
 
 const modalStyleDefaults = {
   content: {
@@ -32,12 +32,12 @@ export default class Viewer extends PureComponent {
     this.state = {
       disabled: false,
     };
-    // this.expandModeData = getImagesData(this.props.initialState);
+    this.expandModeData = getImagesData(this.props.initialState);
   }
 
   componentDidUpdate(prevProps) {
     if (prevProps.initialState !== this.props.initialState) {
-      // this.expandModeData = getImagesData(this.props.initialState);
+      this.expandModeData = getImagesData(this.props.initialState);
     }
   }
 
@@ -49,17 +49,17 @@ export default class Viewer extends PureComponent {
   };
 
   helpers = {
-    // onExpand: (entityIndex, innerIndex = 0) => {
-    //   //galleries can have an innerIndex (i.e. second image will have innerIndex=1)
-    //   this.setState({
-    //     expendModeIsOpen: true,
-    //     expandModeIndex: this.expandModeData.imageMap[entityIndex] + innerIndex,
-    //   });
-    // },
+    onExpand: (entityIndex, innerIndex = 0) => {
+      //galleries have an innerIndex (i.e. second image will have innerIndex=1)
+      this.setState({
+        expendModeIsOpen: true,
+        expandModeIndex: this.expandModeData.imageMap[entityIndex] + innerIndex,
+      });
+    },
   };
 
   render() {
-    // const { expendModeIsOpen, expandModeIndex } = this.state;
+    const { expendModeIsOpen, expandModeIndex } = this.state;
     return (
       <div id="rich-content-viewer" className="viewer">
         <RichContentViewer
@@ -84,14 +84,14 @@ export default class Viewer extends PureComponent {
         >
           <RichContentModal {...this.state.modalProps} />
         </ReactModal>
-        {/* {!isSSR() && (
+        {!isSSR() && (
           <Fullscreen
             isOpen={expendModeIsOpen}
             images={this.expandModeData.images}
             onClose={() => this.setState({ expendModeIsOpen: false })}
             index={expandModeIndex}
           />
-        )} */}
+        )}
       </div>
     );
   }