+++ packages/plugin-image/web/src/image-viewer.jsx
@@ -42,7 +42,7 @@ class ImageViewer extends React.Component {
   getImageUrl(src) {
     const { helpers, shouldRenderOptimizedImages } = this.context || {};
 
-    if (!src && helpers && helpers.handleFileSelection) {
+    if (!src && helpers?.handleFileSelection) {
       return null;
     }