+++ packages/viewer/web/src/RichContentViewer.jsx
@@ -44,6 +44,7 @@ class RichContentViewer extends Component {
       locale,
       disabled,
       shouldRenderOptimizedImages,
+      siteDomain,
     } = this.props;
     return {
       t,
@@ -56,6 +57,7 @@ class RichContentViewer extends Component {
       locale,
       disabled,
       shouldRenderOptimizedImages,
+      siteDomain,
     };
   };
 
@@ -127,6 +129,7 @@ RichContentViewer.propTypes = {
   textDirection: PropTypes.oneOf(['rtl', 'ltr']),
   disabled: PropTypes.bool,
   shouldRenderOptimizedImages: PropTypes.bool,
+  siteDomain: PropTypes.string,
 };
 
 RichContentViewer.defaultProps = {