+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -72,6 +72,7 @@ class RichContentEditor extends Component {
       isMobile,
       shouldRenderOptimizedImages,
       initialIntent,
+      siteDomain,
     } = this.props;
     this.contextualData = {
       theme,
@@ -87,6 +88,7 @@ class RichContentEditor extends Component {
       languageDir: getLangDir(locale),
       shouldRenderOptimizedImages,
       initialIntent,
+      siteDomain,
     };
   };
 
@@ -476,6 +478,7 @@ RichContentEditor.propTypes = {
   shouldRenderOptimizedImages: PropTypes.bool,
   onAtomicBlockFocus: PropTypes.func,
   initialIntent: PropTypes.string,
+  siteDomain: PropTypes.string,
 };
 
 RichContentEditor.defaultProps = {