+++ CHANGELOG.md
@@ -20,6 +20,8 @@
   ### :bug: Bug Fix
   - `common`
     - [#458](https://github.com/wix-incubator/rich-content/pull/458) normalization: sparse entity map keys handled correctly
+  - `fullscreen`
+    - [#457](https://github.com/wix-incubator/rich-content/pull/457) fix fullscreen
   ### :book: Documentation
   - `preview`
     - [#435](https://github.com/wix-incubator/rich-content/pull/435) [documentation](./docs/rich-content-preview.md) added
@@ -45,7 +47,7 @@
     - [#438](https://github.com/wix-incubator/rich-content/pull/438) container alignment
   - `viewer`
     - [#442](https://github.com/wix-incubator/rich-content/pull/442) html plugin alignment
-  - editor
+  - `editor`
     - [#448](https://github.com/wix-incubator/rich-content/pull/448) placeholder style fixed
   ### :house: Internal
   - `gallery`