+++ CHANGELOG.md
@@ -40,6 +40,8 @@
     - [#444](https://github.com/wix-incubator/rich-content/pull/444) file upload support
     - [#449](https://github.com/wix-incubator/rich-content/pull/449) force video mime type for uploads
     - [#451](https://github.com/wix-incubator/rich-content/pull/451) uses unique file input id
+  - `html`
+    - [#456](https://github.com/wix-incubator/rich-content/pull/456) adsense support
   ### :bug: Bug Fix
   - `common`
     - [#454](https://github.com/wix-incubator/rich-content/pull/454) link panel not loading link data after page refresh