+++ CHANGELOG.md
@@ -17,6 +17,9 @@
   ### :bug: Bug Fix
   - `fullscreen`
     - [#465](https://github.com/wix-incubator/rich-content/pull/465) fix fullscreen - getImagesData not working correctly
+  ### :house: Internal
+  - `gallery`
+    - [#467](https://github.com/wix-incubator/rich-content/pull/467) `pro-gallery` updated to `v1.5.33`
 </details>
 <hr/>